move parsers and json utilities to a separate module
